fix building sync transaction timeout, fix building external duplicate

+++ complitex/complitex-common/src/main/java/org/complitex/common/strategy/DomainObjectStrategy.java
@@ -281,7 +281,7 @@ public abstract class DomainObjectStrategy extends AbstractBean implements IStra
 
     @Override
     public Long getObjectId(String externalId) {
-        return (Long) sqlSession().selectOne(NS + ".selectDomainObjectId", ImmutableMap.of("entityName", getEntityName(),
+        return sqlSession().selectOne(NS + ".selectDomainObjectId", ImmutableMap.of("entityName", getEntityName(),
                 "externalId", externalId));
     }
